# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Fintech Risk Monitoring API","version":"0.1.0"},"paths":{"/api/v1/businesses":{"post":{"tags":["businesses"],"summary":"Create Business Route","operationId":"create_business_route_api_v1_businesses_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["businesses"],"summary":"List Businesses Route","operationId":"list_businesses_route_api_v1_businesses_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Case-insensitive substring match","title":"Name"},"description":"Case-insensitive substring match"},{"name":"risk_level","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/RiskLevel"},{"type":"null"}],"description":"Filter by latest completed evaluation risk factor (low, medium, high)","title":"Risk Level"},"description":"Filter by latest completed evaluation risk factor (low, medium, high)"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessListPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/{business_id}/risk-history":{"get":{"tags":["businesses"],"summary":"Risk History","operationId":"risk_history_api_v1_businesses__business_id__risk_history_get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Business Id"}},{"name":"evaluation_status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/RiskEvaluationStatus"},{"type":"null"}],"title":"Evaluation Status"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskHistoryPage"}}}},"404":{"description":"Business not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/{business_id}/evaluate":{"post":{"tags":["businesses"],"summary":"Trigger Evaluate","operationId":"trigger_evaluate_api_v1_businesses__business_id__evaluate_post","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Business Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TriggerEvaluateResponse"}}}},"404":{"description":"Business not found"},"409":{"description":"Risk evaluation already in progress"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/{business_id}":{"get":{"tags":["businesses"],"summary":"Get Business Detail","operationId":"get_business_detail_api_v1_businesses__business_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Business Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusinessDetailRead"}}}},"404":{"description":"Business not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"BusinessCreate":{"properties":{"name":{"type":"string","title":"Name"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"}},"type":"object","required":["name"],"title":"BusinessCreate"},"BusinessDetailRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"latest_completed_risk":{"anyOf":[{"$ref":"#/components/schemas/LatestCompletedRisk"},{"type":"null"}]},"pending_evaluation":{"anyOf":[{"$ref":"#/components/schemas/PendingEvaluationSummary"},{"type":"null"}]}},"type":"object","required":["id","name","industry","created_at"],"title":"BusinessDetailRead"},"BusinessListPage":{"properties":{"items":{"items":{"$ref":"#/components/schemas/BusinessRead"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["items","total","skip","limit"],"title":"BusinessListPage"},"BusinessRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","industry","created_at"],"title":"BusinessRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LatestCompletedRisk":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"score":{"type":"number","title":"Score"},"risk_level":{"type":"string","title":"Risk Level"},"requested_at":{"type":"string","format":"date-time","title":"Requested At"},"completed_at":{"type":"string","format":"date-time","title":"Completed At"}},"type":"object","required":["id","score","risk_level","requested_at","completed_at"],"title":"LatestCompletedRisk"},"PendingEvaluationSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"requested_at":{"type":"string","format":"date-time","title":"Requested At"}},"type":"object","required":["id","requested_at"],"title":"PendingEvaluationSummary"},"RiskEvaluationRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"business_id":{"type":"string","format":"uuid","title":"Business Id"},"status":{"type":"string","title":"Status"},"requested_at":{"type":"string","format":"date-time","title":"Requested At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Score"},"risk_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Risk Level"},"details":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Details"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["id","business_id","status","requested_at","completed_at","score","risk_level","details","error_message"],"title":"RiskEvaluationRead"},"RiskEvaluationStatus":{"type":"string","enum":["pending","in_progress","completed","failed"],"title":"RiskEvaluationStatus"},"RiskHistoryPage":{"properties":{"items":{"items":{"$ref":"#/components/schemas/RiskEvaluationRead"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["items","total","skip","limit"],"title":"RiskHistoryPage"},"RiskLevel":{"type":"string","enum":["low","medium","high"],"title":"RiskLevel","description":"Risk factor tier from latest completed evaluation."},"TriggerEvaluateResponse":{"properties":{"evaluation_id":{"type":"string","format":"uuid","title":"Evaluation Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["evaluation_id","status"],"title":"TriggerEvaluateResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}